All DTOs class moved to shopping-client
