+++ shopping-client/src/test/java/br/com/rocha/shoppingclient/dto/CategoryDTO.java
@@ -1,0 +1,27 @@
+package br.com.rocha.shoppingclient.dto;
+
+import javax.validation.constraints.NotNull;
+
+public class CategoryDTO {
+
+    @NotNull
+    private Long id;
+    private String name;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+}
