+++ shopping-client/src/test/java/br/com/rocha/shoppingclient/dto/ShopReportDTO.java
@@ -1,29 +1,36 @@
-package br.com.rocha.shoppingapi.dto;
+package br.com.rocha.shoppingclient.dto;
 
 public class ShopReportDTO {
-    
+
     private Integer count;
     private Double total;
     private Double mean;
-    
+
+    public ShopReportDTO() {
+    }
+
     public Integer getCount() {
         return count;
     }
+
     public void setCount(Integer count) {
         this.count = count;
     }
+
     public Double getTotal() {
         return total;
     }
+
     public void setTotal(Double total) {
         this.total = total;
     }
+
     public Double getMean() {
         return mean;
     }
+
     public void setMean(Double mean) {
         this.mean = mean;
     }
 
-    
 }
